+++ Source/HtmlRenderer/Core/Dom/CssBoxFrame.cs
@@ -413,8 +413,8 @@ namespace HtmlRenderer.Core.Dom
             tmpRect.Offset(offset);
             tmpRect.Height -= ActualBorderTopWidth + ActualBorderBottomWidth + ActualPaddingTop + ActualPaddingBottom;
             tmpRect.Y += ActualBorderTopWidth + ActualPaddingTop;
-            tmpRect.X = (float)Math.Floor(tmpRect.X);
-            tmpRect.Y = (float)Math.Floor(tmpRect.Y);
+            tmpRect.X = (double)Math.Floor(tmpRect.X);
+            tmpRect.Y = (double)Math.Floor(tmpRect.Y);
             var rect = tmpRect;
 
             DrawImage(g, offset, rect);
